Moved shared logic to utils file. Also added test sample and output snaphot.
